Dev (#9)
* Refactored components to use Material-UI instead of Bootstrap. Started reworking Workmap UI

* Added UI for creating/editing workmap items

* Made workmap items draggble within the container

* Added feature to create paths (draw arrows) between workmap items

* Added functionality to Workmap, can delete and edit path start/end date, and set focus on items

* Updated Workmap UI

* Fixed Home page's column centering. Fixed Workmap position issue by using px instead of rem

* Refactored some firebase code. Improved Workmap interaction

+++ src/components/Workmap/Workmap.jsx
@@ -5,7 +5,7 @@ import WorkmapModal from './WorkmapModal';
 import WorkmapPath from './WorkmapPath';
 import PlainDraggable from 'plain-draggable/plain-draggable.min';
 import { AuthContext } from '../../auth';
-import firebaseApp from '../../firebase';
+import { firestore } from '../../firebase';
 import AddCircleIcon from '@material-ui/icons/AddCircle';
 import { DialogContent, Modal } from '@material-ui/core';
 
@@ -23,7 +23,7 @@ export default function Workmap() {
     const saveItem = (name, abbrev, due, description) => {
         if (!user) return;
         // edit existing workmap item
-        const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
+        const itemsRef = firestore.collection("workmap/" + user.uid + "/items");
         if (currItem)
             return itemsRef.doc(currItem.id).update({
                 name: name.length > 0 ? name : "Task",
@@ -46,8 +46,9 @@ export default function Workmap() {
 
     // handles setting focus on existing workmap items
     const setItemFocus = (itemId, focus) => {
-        const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
-        itemsRef.doc(itemId).update({
+        if (!user) return;
+        const itemsRef = firestore.collection("workmap/" + user.uid + "/items");
+        return itemsRef.doc(itemId).update({
             focus: focus
         });
     };
@@ -55,14 +56,12 @@ export default function Workmap() {
     // handles deleting an existing workmap item
     const deleteItem = () => {
         if (!user) return;
-
-        const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
+        const itemsRef = firestore.collection("workmap/" + user.uid + "/items");
         return itemsRef.doc(currItem.id).delete()
             .then(() => {
                 const deletePaths = pathList.filter((path) => path.from === currItem.id || path.to === currItem.id);
-
                 const deletePromises = [];
-                const pathsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/paths");
+                const pathsRef = firestore.collection("workmap/" + user.uid + "/paths");
                 for (let i = 0; i < deletePaths.length; i++) {
                     deletePromises.push(pathsRef.doc(deletePaths[i].id).delete());
                 }
@@ -72,11 +71,11 @@ export default function Workmap() {
 
     // handles creating a new workmap path
     const newPath = (fromId, toId) => {
-        // path already exists, return
-        if (pathList.filter((path) => path.from === fromId && path.to === toId).length > 0) return;
+        if (!user) return null;
+        if (pathList.filter((path) => path.from === fromId && path.to === toId).length > 0) return null;
         
-        const pathsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/paths")
-        pathsRef.add({
+        const pathsRef = firestore.collection("workmap/" + user.uid + "/paths")
+        return pathsRef.add({
             from: fromId,
             to: toId,
             startDate: null,
@@ -86,7 +85,8 @@ export default function Workmap() {
 
     // handles updating the start and end date of an existing workmap path
     const updatePath = (pathId, startDate, endDate) => {
-        const pathsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/paths");
+        if (!user) return;
+        const pathsRef = firestore.collection("workmap/" + user.uid + "/paths");
         return pathsRef.doc(pathId).update({
             startDate: startDate,
             endDate: endDate
@@ -95,14 +95,15 @@ export default function Workmap() {
 
     // handles deleting an existing workmap path
     const deletePath = (pathId) => {
-        const pathsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/paths");
+        if (!user) return;
+        const pathsRef = firestore.collection("workmap/" + user.uid + "/paths");
         return pathsRef.doc(pathId).delete();
     };
 
     // setup listener for the current user's workmap items and paths
     useEffect(() => {
         if (!user) return;
-        const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
+        const itemsRef = firestore.collection("workmap/" + user.uid + "/items");
         const unsubItem = itemsRef.onSnapshot((qSnapshot) => {
             const itemList = [];
             qSnapshot.forEach((doc) => {
@@ -120,7 +121,7 @@ export default function Workmap() {
             setItemList(itemList);
         });
         
-        const pathsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/paths");
+        const pathsRef = firestore.collection("workmap/" + user.uid + "/paths");
         const unsubPath = pathsRef.onSnapshot((qSnapshot) => {
             const pathList = [];
             qSnapshot.forEach((doc) => {
@@ -144,14 +145,14 @@ export default function Workmap() {
     // make workmap items draggable
     useEffect(() => {
         if (!user) return;
-        const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
+        const itemsRef = firestore.collection("workmap/" + user.uid + "/items");
         itemList.forEach((item) => {
             const domItem = document.getElementById(item.id);
             new PlainDraggable(domItem, {
                 autoScroll: true,
                 leftTop: true,
-                left: item.x + 23, // 21px offset from .workmap-container left padding (16px) + .workmap-content border (7px)
-                top: item.y + 185, // 178px offset from .topbar-container (90px) + .workmap-header (88px) + .workmap-content border (7px)
+                left: item.x + 23, // 23px offset from .workmap-container left padding (16px) + .workmap-content left border (7px)
+                top: item.y + 185, // 185px offset from .topbar-container (90px) + .workmap-header (88px) + .workmap-content top border (7px)
                 onDragEnd: () => {
                     itemsRef.doc(item.id).update({
                         x: Math.round(domItem.style.left.slice(0, -2)),
@@ -160,7 +161,7 @@ export default function Workmap() {
                 }
             });
         });
-    }, [itemList, pathList, user]);
+    }, [itemList, user]);
 
     return (
         <div className="workmap-container">
